Split chat tool execution helpers
Decompose the chat delegate tool execution module into the\nrequested preflight, execution, postflight, and recording\nsubmodules while preserving the existing exported entry points.\n\nAlso apply the remaining verified follow-ups in the affected\ndocs, session tests, thread control helpers, and JobContext\nrollback visibility so the branch reflects the current review\nstate cleanly.
